DTL new-version/backend/requirements.txt
fastapi
uvicorn[standard]
newspaper3k
transformers
torch
feedparser
requests
lxml
beautifulsoup4
